link.h wrong type
Change to unsigned incorrectly increased size.

e88a186fc38c045e4abcbc3c3591a61aa6b6eae9

+++ modules/database/src/ioc/dbStatic/link.h
@@ -193,7 +193,7 @@ struct lset;
 struct link {
     struct dbCommon *precord;   /* Pointer to record owning link */
     short type;
-    unsigned flags;
+    unsigned short flags;
     struct lset *lset;
     char *text;             /* Raw link text */
     union value value;
